tests/wpt/tests/xhr/sync-xhr-supported-by-permissions-policy.tentative.html

<!DOCTYPE html>
<title>Test that sync-xhr is advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-permissions-policy/#dom-permissionspolicy-features">
<link rel="help" href="https://xhr.spec.whatwg.org/#permissions-policy-integration">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('sync-xhr', document.permissionsPolicy.features());
}, 'document.permissionsPolicy.features should advertise sync-xhr.');
</script>
